requirements.txt
Django
gevent
envdir
djangorestframework
django-webpack-loader